Extract label template functions into a label Jinja2 extension
