+++ homeassistant/helpers/template/extensions/__init__.py
@@ -3,6 +3,7 @@
 from .base64 import Base64Extension
 from .collection import CollectionExtension
 from .crypto import CryptoExtension
+from .labels import LabelExtension
 from .math import MathExtension
 from .regex import RegexExtension
 from .string import StringExtension
@@ -11,6 +12,7 @@ __all__ = [
     "Base64Extension",
     "CollectionExtension",
     "CryptoExtension",
+    "LabelExtension",
     "MathExtension",
     "RegexExtension",
     "StringExtension",
